Further simplification and refactor.
Finish and abort now only receive the result/cause because we no longer
persist the received message but rather the fact that it finished or
aborted, and use the _expectedID directly as we know it must be correct
since matchId returned true.
Logs are now uniform and always log the orchestrator name, and the task
name and color when it makes sense.
State is now sealed and the user cannot change it. This change of
approach, ie, not letting the user change the state and add more fields
to it, it what allowed us to greatly simplify finish and abort.
SimpleMessage now only receives and id since the text was never used.